Modified for Android API 8 and much more...

+++ app/src/kiv/janecekz/ma/IControlable.java
@@ -44,5 +44,5 @@ public interface IControlable {
      * @param y
      *            Y coordinate.
      */
-    public void onPositionChange(TouchControl t, float x, float y);
+    public void onPositionChange(TouchControl t, int x, int y);
 }
